// packages/ts/src/types/ast.ts
export interface AnyKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.AnyKeyword;
	readonly parent: TypeNodeParent;
}
// ...
export interface ArrowFunction extends ts.Node {
	readonly asteriskToken?: AsteriskToken | undefined;
	readonly body: ConciseBody;
	readonly equalsGreaterThanToken: EqualsGreaterThanToken;
	readonly exclamationToken?: ExclamationToken | undefined;
	readonly kind: ts.SyntaxKind.ArrowFunction;
	readonly modifiers?: ts.NodeArray<Modifier>;
	readonly name: never;
	readonly parameters: ts.NodeArray<ParameterDeclaration>;
	readonly parent: ExpressionParent;
	readonly questionToken?: QuestionToken | undefined;
	readonly type?: TypeNode | undefined;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration> | undefined;
}
export interface AsExpression extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.AsExpression;
	readonly parent: ExpressionParent;
	readonly type: TypeNode;
}
// ...
export interface AwaitExpression extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.AwaitExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
// ...
export interface BinaryExpression extends ts.Node {
	readonly kind: ts.SyntaxKind.BinaryExpression;
	readonly left: Expression;
	readonly operatorToken: BinaryOperatorToken;
	readonly parent: ExpressionParent;
	readonly right: Expression;
}
// ...
export interface CallExpression extends ts.Node {
	readonly arguments: ts.NodeArray<Expression>;
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.CallExpression;
	readonly parent: LeftHandSideExpressionParent;
	readonly questionDotToken?: QuestionDotToken;
	readonly typeArguments?: ts.NodeArray<TypeNode>;
}
// ...
export interface CommaListExpression extends ts.Node {
	readonly elements: ts.NodeArray<Expression>;
	readonly kind: ts.SyntaxKind.CommaListExpression;
	readonly parent: ExpressionParent;
}
// ...
export interface ConditionalExpression extends ts.Node {
	readonly colonToken: ColonToken;
	readonly condition: Expression;
	readonly kind: ts.SyntaxKind.ConditionalExpression;
	readonly parent: ExpressionParent;
	readonly questionToken: QuestionToken;
	readonly whenFalse: Expression;
	readonly whenTrue: Expression;
}
// ...
export interface Decorator extends ts.Node {
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.Decorator;
	readonly parent:
		| AccessorDeclaration
		| ClassDeclaration
		| ClassExpression
		| ConstructorDeclaration
		| EnumDeclaration
		| ExportAssignment
		| ExportDeclaration
		| FunctionDeclaration
		| ImportDeclaration
		| ImportEqualsDeclaration
		| IndexSignatureDeclaration
		| InterfaceDeclaration
		| JSDocNamespaceDeclaration
		| MethodDeclaration
		| ModuleDeclaration
		| NamespaceDeclaration
		| ParameterDeclaration
		| PropertyDeclaration
		| TypeAliasDeclaration
		| VariableStatement;
}
export interface DefaultClause extends ts.Node {
	readonly kind: ts.SyntaxKind.DefaultClause;
	readonly parent: CaseBlock;
	readonly statements: ts.NodeArray<Statement>;
}
export interface DefaultKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.DefaultKeyword;
	readonly parent: ModifierParent;
}
export interface DeleteExpression extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.DeleteExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
export interface DoStatement extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.DoStatement;
	readonly parent: StatementParent;
	readonly statement: Statement;
}
export interface DotDotDotToken extends ts.Node {
	readonly kind: ts.SyntaxKind.DotDotDotToken;
	readonly parent: BindingElement | ParameterDeclaration;
}
export interface ElementAccessExpression extends ts.Node {
	readonly argumentExpression: Expression;
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.ElementAccessExpression;
	readonly parent: LeftHandSideExpressionParent;
	readonly questionDotToken?: QuestionDotToken;
}
// ...
export type Expression =
	| ArrowFunction
	| AsExpression
	| BinaryExpression
	| CommaListExpression
	| ConditionalExpression
	| JsxClosingFragment
	| JsxExpression
	| JsxOpeningElement
	| JsxOpeningFragment
	| OmittedExpression
	| SatisfiesExpression
	| SpreadElement
	| SyntheticExpression
	| UnaryExpression
	| YieldExpression;
// ...
export interface JsxClosingFragment extends ts.Node {
	readonly kind: ts.SyntaxKind.JsxClosingFragment;
	readonly parent: ExpressionParent | JsxFragment;
}
// ...
export interface JsxExpression extends ts.Node {
	readonly dotDotDotToken?: Token<ts.SyntaxKind.DotDotDotToken, JsxExpression>;
	readonly expression?: Expression;
	readonly kind: ts.SyntaxKind.JsxExpression;
	readonly parent:
		| ArrayLiteralExpression
		| ArrowFunction
		| AsExpression
		| BinaryExpression
		| BindingElement
		| CallExpression
		| CaseClause
		| CommaListExpression
		| ComputedPropertyName
		| ConditionalExpression
		| ElementAccessExpression
		| EnumMember
		| ExportAssignment
		| ExportDeclaration
		| ExpressionStatement
		| ExternalModuleReference
		| IfStatement
		| ImportAttribute
		| ImportDeclaration
		| IterationStatement
		| JsxAttributeLike
		| JsxElement
		| JsxExpression
		| JsxFragment
		| NewExpression
		| NonNullExpression
		| ParameterDeclaration
		| ParenthesizedExpression
		| PartiallyEmittedExpression
		| PropertyAssignment
		| PropertyDeclaration
		| ReturnStatement
		| SatisfiesExpression
		| ShorthandPropertyAssignment
		| SpreadAssignment
		| SpreadElement
		| SwitchStatement
		| TemplateSpan
		| ThrowStatement
		| TypeParameterDeclaration
		| VariableDeclaration
		| WithStatement
		| YieldExpression;
}
// ...
export interface JsxOpeningElement extends ts.Node {
	readonly attributes: JsxAttributes;
	readonly kind: ts.SyntaxKind.JsxOpeningElement;
	readonly parent: ExpressionParent | JsxElement;
	readonly tagName: JsxTagNameExpression;
	readonly typeArguments?: ts.NodeArray<TypeNode>;
}
export interface JsxOpeningFragment extends ts.Node {
	readonly kind: ts.SyntaxKind.JsxOpeningFragment;
	readonly parent: ExpressionParent | JsxFragment;
}
// ...
export type LeftHandSideExpression =
	| CallExpression
	| MemberExpression
	| NonNullExpression
	| PartiallyEmittedExpression;
// ...
export type MemberExpression =
	| ElementAccessExpression
	| ExpressionWithTypeArguments
	| PrimaryExpression
	| PropertyAccessExpression
	| TaggedTemplateExpression;
// ...
export interface NonNullExpression extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.NonNullExpression;
	readonly parent: LeftHandSideExpressionParent;
}
// ...
export interface OmittedExpression extends ts.Node {
	readonly kind: ts.SyntaxKind.OmittedExpression;
	readonly parent: ArrayBindingPattern | ExpressionParent;
}
// ...
export interface PartiallyEmittedExpression extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.PartiallyEmittedExpression;
	readonly parent: LeftHandSideExpressionParent;
}
export interface PlusToken extends ts.Node {
	readonly kind: ts.SyntaxKind.PlusToken;
	readonly parent: MappedTypeNode;
}
export interface PostfixUnaryExpression extends ts.Node {
	readonly kind: ts.SyntaxKind.PostfixUnaryExpression;
	readonly operand: LeftHandSideExpression;
	readonly operator: PostfixUnaryOperator;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
export type PostfixUnaryOperator =
	| ts.SyntaxKind.MinusMinusToken
	| ts.SyntaxKind.PlusPlusToken;
export interface PrefixUnaryExpression extends ts.Node {
	readonly kind: ts.SyntaxKind.PrefixUnaryExpression;
	readonly operand: UnaryExpression;
	readonly operator: PrefixUnaryOperator;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| LiteralTypeNode
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
// ...
export interface PropertyAccessExpression extends ts.Node {
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.PropertyAccessExpression;
	readonly name: MemberName;
	readonly parent: LeftHandSideExpressionParent;
	readonly questionDotToken?: QuestionDotToken;
}
// ...
export interface QuestionDotToken extends ts.Node {
	readonly kind: ts.SyntaxKind.QuestionDotToken;
	readonly parent:
		| CallExpression
		| ElementAccessExpression
		| JsxTagNamePropertyAccess
		| PropertyAccessExpression;
}
// ...
export interface SatisfiesExpression extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.SatisfiesExpression;
	readonly parent: ExpressionParent;
	readonly type: TypeNode;
}
// ...
export interface SpreadElement extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.SpreadElement;
	readonly parent: ExpressionParent;
}
// ...
export interface SyntheticExpression extends ts.Node {
	readonly isSpread: boolean;
	readonly kind: ts.SyntaxKind.SyntheticExpression;
	readonly parent: ExpressionParent;
	readonly tupleNameSource?: NamedTupleMember | ParameterDeclaration;
	readonly type: ts.Type;
}
export interface TaggedTemplateExpression extends ts.Node {
	readonly kind: ts.SyntaxKind.TaggedTemplateExpression;
	readonly parent: LeftHandSideExpressionParent;
	readonly tag: LeftHandSideExpression;
	readonly template: TemplateLiteral;
	readonly typeArguments?: ts.NodeArray<TypeNode>;
}
// ...
export type TemplateLiteral =
	| NoSubstitutionTemplateLiteral
	| TemplateExpression;
// ...
export interface TypeAssertion extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.TypeAssertionExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
	readonly type: TypeNode;
}
// ...
export type TypeNode =
	| AnyKeyword
	| ArrayTypeNode
	| BigIntKeyword
	| BooleanKeyword
	| ConditionalTypeNode
	| FunctionOrConstructorTypeNodeBase
	| IndexedAccessTypeNode
	| InferTypeNode
	| IntersectionTypeNode
	| IntrinsicKeyword
	| JSDocType
	| JSDocTypeExpression
	| LiteralTypeNode
	| MappedTypeNode
	| NamedTupleMember
	| NeverKeyword
	| NodeWithTypeArguments
	| NumberKeyword
	| ObjectKeyword
	| OptionalTypeNode
	| ParenthesizedTypeNode
	| RestTypeNode
	| StringKeyword
	| SymbolKeyword
	| TemplateLiteralTypeNode
	| TemplateLiteralTypeSpan
	| ThisTypeNode
	| TupleTypeNode
	| TypeLiteralNode
	| TypeOperatorNode
	| TypePredicateNode
	| UndefinedKeyword
	| UnionTypeNode
	| UnknownKeyword
	| VoidKeyword;
export interface TypeOfExpression extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.TypeOfExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
// ...
export type UnaryExpression =
	| AwaitExpression
	| DeleteExpression
	| TypeAssertion
	| TypeOfExpression
	| UpdateExpression
	| VoidExpression;
// ...
export interface VoidExpression extends ts.Node {
	readonly expression: UnaryExpression;
	readonly kind: ts.SyntaxKind.VoidExpression;
	readonly parent:
		| AwaitExpression
		| DeleteExpression
		| ExpressionParent
		| PrefixUnaryExpression
		| TypeAssertion
		| TypeOfExpression
		| VoidExpression;
}
// ...
export interface YieldExpression extends ts.Node {
	readonly asteriskToken?: AsteriskToken;
	readonly expression?: Expression;
	readonly kind: ts.SyntaxKind.YieldExpression;
	readonly parent: ExpressionParent;
}
// ...
export type ExpressionParent =
	| ArrayLiteralExpression
	| ArrowFunction
	| AsExpression
	| BinaryExpression
	| BindingElement
	| CallExpression
	| CaseClause
	| CommaListExpression
	| ComputedPropertyName
	| ConditionalExpression
	| ElementAccessExpression
	| EnumMember
	| ExportAssignment
	| ExportDeclaration
	| ExpressionStatement
	| ExternalModuleReference
	| IfStatement
	| ImportAttribute
	| ImportDeclaration
	| IterationStatement
	| JsxExpression
	| JsxSpreadAttribute
	| NewExpression
	| NonNullExpression
	| ParameterDeclaration
	| ParenthesizedExpression
	| PartiallyEmittedExpression
	| PropertyAssignment
	| PropertyDeclaration
	| ReturnStatement
	| SatisfiesExpression
	| ShorthandPropertyAssignment
	| SpreadAssignment
	| SpreadElement
	| SwitchStatement
	| TemplateSpan
	| ThrowStatement
	| TypeParameterDeclaration
	| VariableDeclaration
	| WithStatement
	| YieldExpression;
export type LeftHandSideExpressionParent =
	| AwaitExpression
	| Decorator
	| DeleteExpression
	| ExpressionParent
	| ExpressionWithTypeArguments
	| PostfixUnaryExpression
	| PrefixUnaryExpression
	| PropertyAccessExpression
	| TaggedTemplateExpression
	| TypeAssertion
	| TypeOfExpression
	| VoidExpression;